git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@1722 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-maven-web/src/main/java/de/cebitec/mgx/web/StatisticsBean.java
@@ -40,9 +40,9 @@ public class StatisticsBean {
     @Consumes("application/x-protobuf")
     @Produces("application/x-protobuf")
     public PointDTOList create(MGXLongList dto) {
-        long[] data = new long[dto.getLongCount()];
+        double[] data = new double[dto.getLongCount()];
         for (int i = 0; i < dto.getLongCount(); i++) {
-            data[i] = dto.getLong(i).getValue();
+            data[i] = (double) dto.getLong(i).getValue();
         }
         AutoCloseableIterator<Point> ret;
         try {
